+++ doc-site/.vitepress/theme/components/TypographyTable.vue
@@ -106,9 +106,9 @@ watch(
 
 watch(
   () => [
-    cssTokenState.headers,
-    cssTokenState.subHeaders,
-    cssTokenState.ingress,
+    cssTokenState.headings,
+    cssTokenState.subHeadings,
+    cssTokenState.lead,
     cssTokenState.body,
     cssTokenState.bodyCompact,
     cssTokenState.detailText,
@@ -117,16 +117,16 @@ watch(
   () => {
     switch (props.tableContentType) {
       case 'default':
-        tableContent.value = cssTokenState.headers;
+        tableContent.value = cssTokenState.headings;
         break;
-      case 'headers':
-        tableContent.value = cssTokenState.headers;
+      case 'headings':
+        tableContent.value = cssTokenState.headings;
         break;
-      case 'subheaders':
-        tableContent.value = cssTokenState.subHeaders;
+      case 'subheadings':
+        tableContent.value = cssTokenState.subHeadings;
         break;
-      case 'ingress':
-        tableContent.value = cssTokenState.ingress;
+      case 'lead':
+        tableContent.value = cssTokenState.lead;
         break;
       case 'body':
         tableContent.value = cssTokenState.body;
@@ -159,8 +159,8 @@ watch(
 }
 
 table {
-  --vp-c-divider: var(--grey-200);
-  border: solid 1px var(--grey-200);
+  --vp-c-divider: var(--color-shades-grey-200);
+  border: solid 1px var(--color-shades-grey-200);
 }
 
 .vp-doc tr:nth-child(2n) {
@@ -168,16 +168,16 @@ table {
 }
 
 nve-radio-group {
-  border: solid 1px var(--grey-200);
+  border: solid 1px var(--color-shades-grey-200);
   border-bottom: none;
   border-radius: 3px 3px 0px 0px;
   padding: 10px;
 }
 
 .font-list {
-  border-bottom: solid 1px var(--grey-200);
-  border-left: solid 1px var(--grey-200);
-  border-right: solid 1px var(--grey-200);
+  border-bottom: solid 1px var(--color-shades-grey-200);
+  border-left: solid 1px var(--color-shades-grey-200);
+  border-right: solid 1px var(--color-shades-grey-200);
   border-radius: 0px 0px 3px 3px;
   min-width: 20rem;
   max-width: 100%;
@@ -190,7 +190,7 @@ nve-radio-group {
 }
 
 .font-list--border-top {
-  border-top: solid 1px var(--grey-200);
+  border-top: solid 1px var(--color-shades-grey-200);
 }
 
 .font-list__container table,