+++ src/components/nve-textarea/nve-textarea.styles.ts
@@ -18,17 +18,17 @@ export default css`
   }
 
   .textarea__control {
-    font: var(--body-small);
+    font: var(--typography-body-small);
     box-sizing: border-box;
     padding: var(--sizing-4x-small);
     padding-right: var(--sizing-2x-small); /** trenger padding for å vise ikone så at teksten ikke dekker den */
     border-radius: var(--border-radius-small);
-    border: var(--border-width-default, 1px) solid var(--neutrals-border-default);
+    border: var(--border-width-default, 1px) solid var(--color-neutrals-border-default);
     min-height: var(--sizing-2x-small);
     transition: border var(--transition-time) ease-in-out;
     width: 100%;
     &:hover:not(:disabled) {
-      border-color: var(--neutrals-foreground-primary);
+      border-color: var(--color-neutrals-foreground-primary);
     }
 
     &:focus-visible {
@@ -38,30 +38,30 @@ export default css`
   }
 
   :host([data-user-invalid]) .textarea__control {
-    border-color: var(--feedback-background-emphasized-error);
+    border-color: var(--color-feedback-background-emphasized-error);
   }
 
   :host([disabled]) .textarea__control {
     opacity: 0.38;
-    background: var(--neutrals-background-primary-contrast);
+    background: var(--color-neutrals-background-primary-contrast);
   }
 
   :host([filled]) .textarea__control {
-    background: var(--neutrals-background-primary-contrast);
-    border: var(--border-width-default, 1px) solid var(--neutrals-border-subtle);
+    background: var(--color-neutrals-background-primary-contrast);
+    border: var(--border-width-default, 1px) solid var(--color-neutrals-border-subtle);
 
     &:hover:not(:disabled) {
-      border-color: var(--neutrals-border-default);
+      border-color: var(--color-neutrals-border-default);
     }
   }
 
   :host([filled][data-user-invalid]) .textarea__control {
-    border-color: var(--feedback-background-emphasized-error);
+    border-color: var(--color-feedback-background-emphasized-error);
   }
 
   :host([readonly]) .textarea__control {
-    background: var(--neutrals-background-secondary);
-    color: var(--neutrals-foreground-subtle,);
+    background: var(--color-neutrals-background-secondary);
+    color: var(--color-neutrals-foreground-subtle,);
     border: none;
   }
 
@@ -74,9 +74,9 @@ export default css`
 
   :host([required]) nve-label::after {
     content: var(--textarea-required-content);
-    font: var(--label-x-small-light);
+    font: var(--typography-label-x-small-light);
     margin-left: auto;
-    color: var(--feedback-background-emphasized-error);
+    color: var(--color-feedback-background-emphasized-error);
   }
 
   .textarea__help-text__container {
@@ -86,12 +86,12 @@ export default css`
   }
 
   .textarea__help-text {
-    font: var(--label-x-small-light);
+    font: var(--typography-label-x-small-light);
     color: var(--sl-input-help-text-color);
   }
 
   .textarea__help-text--error {
-    color: var(--feedback-background-emphasized-error);
+    color: var(--color-feedback-background-emphasized-error);
   }
 
   .textarea__icon__container {
@@ -102,7 +102,7 @@ export default css`
     position: absolute;
     left: -24px;
     top: 10px;
-    color: var(--feedback-background-emphasized-error);
+    color: var(--color-feedback-background-emphasized-error);
   }
 
   nve-label {